backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pdfplumber>=0.10.3
google-generativeai>=0.5.0
python-dotenv>=1.0.1
pydantic[email]>=2.7.0
httpx>=0.27.0
